authen authorization, admin

+++ server/controllers/todos.js
@@ -8,7 +8,19 @@ class Controller{
             res.status(200).json(data)
         })
         .catch(err=>{
-            res.status(400).json('data tidak ada')
+            res.status(400).json('internal server error')
+        })
+    }
+
+    static findAllMyTodos(req,res){
+        const username = req.userData.username
+        console.log(username)
+        Todo.findAll({where : {username : username}})
+        .then(data=>{
+            res.status(200).json(data)
+        })
+        .catch(err=>{
+            res.status(400).json('internal server error')
         })
     }
 
@@ -19,12 +31,13 @@ class Controller{
             res.status(200).json(data)
         })
         .catch(err=>{
-            res.status(400).json('data tidak ada')
+            res.status(400).json('internal server error')
         })
     }
 
     static add(req,res){
         const body = {
+            username:req.body.username,
             title:req.body.title,
             description:req.body.description,
             status:req.body.status,
@@ -36,42 +49,68 @@ class Controller{
             res.status(201).json(data)
         })
         .catch(err=>{
-            res.status(200).json({message: `data not found`})
+            res.status(400).json({message: `internal server error`})
         })
     }
 
     static update(req,res){
         const id = req.params.id
-        const body = {
-            title:req.body.title,
-            description:req.body.description,
-            status:req.body.status,
-            due_date:req.body.due_date
-        }
+        const username = req.userData.username
+        const body = req.body
+        let status = true
+            body.username = body.username,
+            body.title = req.body.title,
+            body.description = req.body.description,
+            body.status = req.body.status,
+            body.due_date = req.body.due_date
 
-        Todo.update(body, {where : {id : id}})
+        Todo.findByPk(id)
+        .then((data)=>{
+            if(username.toLowerCase().slice(0,5) == 'admin'){
+                return Todo.update(body, {where : {id : id}})
+            }
+            else if(username !== data.username){
+                status = false
+                res.status(501).json('un - authorized only for admin or user only')
+            }
+            else if(username == data.username){
+                return Todo.update(body, {where : {id : id}})
+            }
+        })
         .then(data=>{
             if(data == 1){
                 res.status(200).json({message: `data updated`})
             }else{
                 res.status(400).json({message: `data not found`})
             }
-            
         })
         .catch(err=>{
-            console.log(err)
+            res.status(400).json({message: `internal server error`})
         })
     }
 
     static delete(req,res){
         const id = req.params.id
-        console.log(id)
-        Todo.destroy({where: {id : id}})
+        const username = req.userData.username
+        let status = true
+        Todo.findByPk(id)
+        .then(data=>{
+            if(username.toLowerCase().slice(0,5) == 'admin'){
+                return Todo.destroy({where: {id : id}})
+            }
+            else if(username !== data.username){
+                status = false
+                res.status(501).json('un - authorized only for admin or user only')
+            }
+            else if(username == data.username){
+                return Todo.destroy({where: {id : id}})
+            }
+        })
         .then(data=>{
             res.status(200).json(data)
         })
         .catch(err=>{
-            res.status(200).json({message: `data not found`})
+            res.status(400).json({message: `data not found`})
         })
     }
